requirements.txt
click==8.1.3
cmake==3.25.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
